# user-microservice/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: User Microservice
  description: |-
    This is the User Microservice for the EasyConf scenario in CSE2115. It is responsible for managing users, events (including their tracks and attendees), and invitations.
  version: 0.0.0
servers:
  - url: http://localhost:8085
    description: Production server for integration.
...
paths:
  /attendee:
    post:
      tags:
        - Event Management
      summary: Create a new attendee.
      description: The invitation system should be used by human users of the API. This endpoint is only designed to be used by other microservices.
      operationId: createAttendee
      requestBody:
        description: Attendee to be created.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Attendee'
      responses:
        '200':
          description: Operation executed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Attendee'
        '401':
          description: Unauthorized access.
        '409':
          description: Attendee already exists.
      security:
        - api_key: [ ]

    get:
      tags:
        - Event Management
        - Track Management
      summary: Returns all attendees, optionally filtered by an event, roles, or track.
      operationId: getInvitesByEventID
      parameters:
        - name: eventID
          required: false
          in: query
          description: Event identifier filter.
          schema:
            type: integer
            format: int64
        - name: roles
          required: false
          in: query
          description: Role filters
          schema:
            type: array
            items:
              $ref: '#/components/schemas/Role'
        - name: trackID
          required: false
          in: query
          description: Track identifier filter
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Attendee'
        '400':
          description: Invalid filters were provided.
        '401':
          description: Unauthorized access.
      security:
        - api_key: []
    put:
      tags:
        - Event Management
        - Track Management
      summary: Updates the role of an attendee in the specified event or track.
      description: This operation can only be performed by someone with an equal or higher role in the given event or track than the attendee's current and target roles.
      operationId: updateAttendee
      requestBody:
        description: The attendee object to be updated.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Attendee'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Attendee'
        '400':
          description: Invalid attendee object was provided.
        '401':
          description: Unauthorized access.
        '404':
          description: No attendee with the provided ID exists.
      security:
        - api_key: [ ]
  /attendee/{attendeeID}:
    get:
      tags:
        - Event Management
      summary: Get an attendee by ID.
      operationId: getAttendeeByID
      parameters:
        - name: attendeeID
          required: true
          in: path
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Attendee'
        '400':
          description: Invalid attendee ID was provided.
        '401':
          description: Unauthorized access.
        '404':
          description: No attendee with the provided ID exists.
      security:
        - api_key: [ ]
    delete:
      tags:
        - Event Management
        - Track Management
      summary: Remove an attendee from a particular event or track.
      description: This operation can only be performed by someone with an equal or higher role in the given event or track than the attendee's role.
      operationId: deleteAttendee
      parameters:
        - name: attendeeID
          required: true
          in: path
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
        '400':
          description: Invalid attendeeID was provided.
        '401':
          description: Unauthorized access.
        '404':
          description: The user specified with the userID never attended the event specified by the eventID.
      security:
        - api_key: [ ]
...
components:
  schemas:
...
    Attendee:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        event_id:
          type: integer
          format: int64
          example: 10
        user_id:
          type: integer
          format: int64
          example: 10
        track_id:
          type: integer
          nullable: true
          format: int64
          example: 10
        role:
          $ref: '#/components/schemas/Role'
...
    Role:
      type: string
      enum:
        - general_chair
        - pc_chair
        - pc_member
        - sub_reviewer
        - author
        - attendee
...
  securitySchemes:
    api_key:
      type: http
      scheme: bearer
      bearerFormat: JWT